// RSCDevStats/ClientApp/src/models/PlayerStats.ts
export interface PlayerStats {
	Name: string
	OnlineId: string
	Team: number
	Score: number
	Goals: number
	Assists: number
	Saves: number
	Shots: number
}

// {
//   "Name": "Vengeance",
//   "Platform": {
//     "Type": "OnlinePlatform",
//     "Value": "OnlinePlatform_Steam"
//   },
//   "OnlineID": 76561198251767546,
//   "Team": 1,
//   "Score": 348,
//   "Goals": 0,
//   "Assists": 2,
//   "Saves": 2,
//   "Shots": 0,
//   "bBot": 0
// }